(BSR)[API] chore: move ubble script in ubble api
