--- api/src/pcapi/scripts/subscription/ubble/__init__.py
@@ -1,1 +1,0 @@
-from .update_pending_ubble_applications import update_pending_ubble_applications
